fix: Issue #694
Add tabs reconsile when merging profile.
